refactor: hide design settings
