+++ src/DonationForms/resources/app/DonationFormApp.tsx
@@ -90,63 +90,6 @@ function AppPreview() {
 
     useEffect(() => {
         subscribeToSettings((settings) => {
-            if (settings['designSettingsSectionStyle']) {
-                updateDesignSettingsClassName(
-                    'givewp-design-settings--section-style',
-                    settings['designSettingsSectionStyle']
-                );
-            }
-
-            if (settings['designSettingsImageUrl']) {
-                root.style.setProperty(
-                    '--givewp-design-settings-background-image',
-                    'url(' + settings['designSettingsImageUrl'] + ')'
-                );
-
-                const style = settings['designSettingsImageStyle'] ? settings['designSettingsImageStyle'] : 'background';
-
-                updateDesignSettingsClassName('givewp-design-settings--image-style', style);
-            }
-
-            if (settings['designSettingsLogoUrl']) {
-                root.style.setProperty(
-                    '--givewp-design-settings-logo',
-                    'url(' + settings['designSettingsLogoUrl'] + ')'
-                );
-                root.classList.add('givewp-design-settings--logo');
-
-                const position = settings['designSettingsLogoPosition'] ? settings['designSettingsLogoPosition'] : 'left';
-                updateDesignSettingsClassName('givewp-design-settings--logo-position', position);
-            }
-
-            if (settings['designSettingsTextFieldStyle']) {
-                updateDesignSettingsClassName(
-                    'givewp-design-settings--textField-style',
-                    settings['designSettingsTextFieldStyle']
-                );
-            }
-
-            if (!settings['designSettingsImageUrl']) {
-                // reset/remove classnames on delete
-                root.style.setProperty('--givewp-design-settings-background-image', '');
-                updateDesignSettingsClassName('givewp-design-settings--image-style', '');
-
-                // reconstruct branding container & logo container
-                root.classList.add('givewp-design-settings--logo');
-                root.style.setProperty(
-                    '--givewp-design-settings-logo',
-                    'url(' + settings['designSettingsLogoUrl'] + ')'
-                );
-                updateDesignSettingsClassName('givewp-design-settings--logo-position', settings['designSettingsLogoPosition']);
-            }
-
-            // reset/remove classnames on delete
-            if (!settings['designSettingsLogoUrl']) {
-                root.style.setProperty('--givewp-design-settings-logo', '');
-                root.classList.remove('givewp-design-settings--logo');
-                updateDesignSettingsClassName('givewp-design-settings--logo-position', '');
-            }
-
             setFormState((prevState) => {
                 return {
                     ...prevState,
@@ -197,15 +140,6 @@ function AppPreview() {
     }, []);
 
     return <App form={formState} />;
-}
-
-function updateDesignSettingsClassName(block, element) {
-    root.classList.forEach((className) => {
-        if (className.startsWith(block + '__')) {
-            root.classList.remove(className);
-        }
-    });
-    root.classList.add(block + '__' + element);
 }
 
 const root = document.getElementById('root-givewp-donation-form');
