+++ src/FormBuilder/resources/js/form-builder/src/settings/design/general-controls/index.tsx
@@ -31,7 +31,6 @@ export default function GeneralControls() {
         >
             <Layout
                 dispatch={dispatch}
-                publishSettings={publishSettings}
                 formDesigns={formDesigns}
                 designId={designId}
             />
