+++ src/FormBuilder/resources/js/form-builder/src/settings/design/general-controls/input-fields/index.tsx
@@ -1,0 +1,54 @@
+import {PanelBody, PanelRow, SelectControl} from '@wordpress/components';
+import {__} from '@wordpress/i18n';
+import {setFormSettings, useFormState} from '@givewp/form-builder/stores/form-state';
+
+/**
+ * @note currently unused
+ * @unreleased
+ */
+export default function InputFields({dispatch, publishSettings}) {
+    const {
+        settings: {
+            designSettingsTextFieldStyle,
+            designSettingsImageStyle,
+            designSettingsImageUrl,
+            designSettingsLogoUrl,
+            designSettingsLogoPosition,
+        },
+    } = useFormState();
+
+    return (
+        <PanelBody title={__('Input Fields', 'give')}>
+            <PanelRow>
+                <SelectControl
+                    label={__('Input Field Style', 'give')}
+                    help={__('Change the design of the input fields for this form.', 'give')}
+                    onChange={(designSettingsTextFieldStyle) => {
+                        dispatch(
+                            setFormSettings({
+                                designSettingsTextFieldStyle,
+                                designSettingsImageStyle,
+                                designSettingsImageUrl,
+                                designSettingsLogoUrl,
+                                designSettingsLogoPosition,
+                            })
+                        );
+                        publishSettings({
+                            designSettingsTextFieldStyle,
+                            designSettingsImageStyle,
+                            designSettingsImageUrl,
+                            designSettingsLogoUrl,
+                            designSettingsLogoPosition,
+                        });
+                    }}
+                    value={designSettingsTextFieldStyle}
+                    options={[
+                        {label: __('Default', 'give'), value: 'default'},
+                        {label: __('Box (inner-label)', 'give'), value: 'box'},
+                        {label: __('Line (inner-label)', 'give'), value: 'line'},
+                    ]}
+                />
+            </PanelRow>
+        </PanelBody>
+    );
+}
